==> src/Batch_API_Method/1_Post_History_FIrst_Version_Datasets.ipynb <==
import pandas as pd

# 1. Load the Datasets and Filter only the Initial Title, Body and Tags

# Step 1: Load the data
df = pd.read_csv('./data/processed/Post_Combined.csv')

# Step 2: Filter only the Initial Title, Body, Tags, Comments and Answers
filter = df['Name'].isin(['Initial Title', 'Initial Body', 'Initial Tags', 'Comments', 'Answers'])
df = df[filter]

# Debug
print(df)

## Grouped by PostID and Extract each row Initial version

# Step 3: Group by PostId
grouped = df.groupby('PostId')

# # Debug: Display the dataframe after group by PostID object
# for post_id, group in grouped:
#     print(f"Post ID: {post_id}")
#     print(group)
#     print("-" * 40)
    
# Step 4: For each PostId, extract Title, Body, and Tags
records = []

for post_id, group in grouped:
    # Dictionary for each row. Has PostID, Initial Title, Initial BOdy,Initial Tag
    post = {
        'PostId': post_id,
        'Title': None,
        'Body': None,
        'Tags': None,
        'Comments': [],
        'Answers': []
    }
    
    # Create List for Comments and Answers so that the 'post' Dictionary could include List
    comments = []
    answers = []

    for row in group.itertuples():
        if row.Name == 'Initial Title':
            post['Title'] = row.Text
        elif row.Name == 'Initial Body':
            post['Body'] = row.Text
        elif row.Name == 'Initial Tags':
            post['Tags'] = row.Text
        elif row.Name == 'Comments':
            comments.append((row.Score, row.Text))
        elif row.Name == 'Answers':
            answers.append((row.Score, row.Text))
            
    # Sort and store top 3 by score
    post['Comments'] = [text for score, text in sorted(comments, reverse=True)[:3]]
    post['Answers'] = [text for score, text in sorted(answers, reverse=True)[:3]]
    
    records.append(post)
    
# Debug
#print(records[0])

# Step 5: Turn records List into DataFrame
final_df = pd.DataFrame(records)

# Debug
from IPython.display import display
display(final_df)

# Step 6: Export to csv file
final_df.to_csv(path_or_buf='./data/processed/PostHistoryFirstVersion.csv', index = False) # False index param means no index for export data
