==> src/Batch_API_Method/3_Post_History_Batch_API.ipynb <==
# 3. Upload the Batch input file for the LLM to process

import pandas as pd
import os
from dotenv import load_dotenv
from openai import OpenAI
import json

# Load the API key from config file
load_dotenv("./config/api_key.env")
# Fetch the API key from the environment
api_key = os.getenv("OPENAI_API_KEY")
client = OpenAI(api_key=api_key)

# Upload the batch input file 
batch_input_file = client.files.create(
    file=open("./data/processed/batch_input.jsonl", "rb"),
    purpose="batch"
)

print(batch_input_file)

# Create the batch for process based on the batch input file ID
batch_input_file_id = batch_input_file.id

client.batches.create(
    input_file_id=batch_input_file_id,
    endpoint="/v1/chat/completions",
    completion_window="24h",
    metadata={
        "description": "Testing Batches of Stackoverflow Questions for OpenAI to process"
    }
)

## Check the status of the Batch and Retrieve the result when done

# Get the status of the batch ( id = batct_123 )
batch = client.batches.retrieve("batch_6821541dc1d881908ddcb473b7a01d13")
print(batch)

# Retrieve the result ( output_file_id='file-123')
file_response = client.files.content("file-P1JCB4sx51YmhXW9ywRY4A")
print(file_response.text)

## Output the response part (Put it in a List, CSV, JSON file,...)

# Full response data as a JSON string
json_data = file_response.text

# Full response data, put each line in a List
json_list = []
for line in json_data.strip().split("\n"):
    if line.strip():
        json_list.append(json.loads(line))  # parse a JSON string and convert it to a Python Dictionary, then append to a List

# (Optional) Put them in a batch output file
with open("./data/processed/batch_output.json", "w", encoding="utf-8") as json_file: 
    for item in json_list:
        json_file.write(json.dumps(item, ensure_ascii = False, indent = 4) + "\n") 


# Debug json List
print(json_list[2]) 

## Get the Response content and get the Title, Body and Tags part only

# List to store the improved version of title, body and tags
new_titles = []
new_bodies = []
new_tags = []

# Parse the response by Title, Body and Tags
for data in json_list:
    # Parse the response
    response = data["response"]["body"]["choices"][0]["message"]["content"]
    
    # Very basic parsing
    title_only = response.split('Title:')[1].split('Body:')[0].strip()
    body_only = response.split('Body:')[1].split('Tags:')[0].strip()
    tags_only = response.split('Tags:')[1].strip()
    
    # append each new Title, Body and Tags in a List
    new_titles.append(title_only)
    new_bodies.append(body_only)
    new_tags.append(tags_only)

# Load the dataset
df = pd.read_csv('./data/processed/PostHistoryFirstVersion.csv')

# print(new_titles)
# print(new_bodies)
# print(new_tags)

# Add new columns to the DataFrame
df['Improved_Title'] = new_titles
df['Improved_Body'] = new_bodies
df['Improved_Tags'] = new_tags

# Save to new CSV
df.to_csv('./data/processed/ImprovedQuestions.csv', index=False)
